requirements.txt
PySide6>=6.6.0
PyYAML>=6.0.1
dnspython>=2.6.0
ping3>=4.0.8
google-genai>=0.5.0
aiohttp>=3.9.3
